Move all Alice and Bob related things into /alice and /bob
