+++ application/comit_node/src/http_api/rfc003/alice_spawner.rs
@@ -8,9 +8,8 @@ use crate::{
         rfc003::{
             alice::SwapRequest,
             events::{LedgerEvents, LqsEvents, LqsEventsForErc20},
-            roles::{Alice, Initiation},
             state_store::{self, StateStore},
-            Ledger, SecretSource,
+            Alice, Initiation, Ledger, SecretSource,
         },
         SwapId,
     },
