+++ application/comit_node/src/swap_protocols/rfc003/alice/communication_events.rs
@@ -6,7 +6,7 @@ use crate::{
             self,
             events::{CommunicationEvents, ResponseFuture, StateMachineResponseFuture},
             ledger::Ledger,
-            roles::Alice,
+            Alice,
         },
     },
 };
