+++ application/comit_node/src/swap_protocols/rfc003/bob/actions/btc_erc20.rs
@@ -1,15 +1,12 @@
 use crate::swap_protocols::{
     ledger::{Bitcoin, Ethereum},
     rfc003::{
-        actions::{
-            bob::{Accept, Decline},
-            ActionKind, Actions,
-        },
         bitcoin,
+        bob::actions::{Accept, Decline},
         ethereum::{self, Erc20Htlc, Htlc},
-        roles::Bob,
         secret::Secret,
         state_machine::*,
+        ActionKind, Actions, Bob,
     },
 };
 use bitcoin_support::{BitcoinQuantity, OutPoint};
