+++ application/comit_node/src/swap_protocols/rfc003/bob/communication_events.rs
@@ -5,7 +5,7 @@ use crate::{
         rfc003::{
             events::{CommunicationEvents, ResponseFuture},
             ledger::Ledger,
-            roles::Bob,
+            Bob,
         },
     },
 };
