+++ application/comit_node/src/swap_protocols/rfc003/bob/handler.rs
@@ -6,12 +6,11 @@ use crate::{
         metadata_store::MetadataStore,
         rfc003::{
             self,
-            bob::SwapRequestKind,
-            events::{BobToAlice, CommunicationEvents, LedgerEvents, LqsEvents, LqsEventsForErc20},
-            roles::Bob,
+            bob::{BobToAlice, SwapRequestKind},
+            events::{CommunicationEvents, LedgerEvents, LqsEvents, LqsEventsForErc20},
             state_machine::*,
             state_store::StateStore,
-            Ledger,
+            Bob, Ledger,
         },
         SwapId,
     },
