+++ application/comit_node/src/swap_protocols/rfc003/mod.rs
@@ -1,14 +1,12 @@
 #[macro_use]
 mod transition_save;
 
-pub mod actions;
 pub mod alice;
 pub mod bitcoin;
 pub mod bob;
 pub mod ethereum;
 pub mod events;
 pub mod find_htlc_location;
-pub mod roles;
 
 pub mod state_machine;
 pub mod state_store;
@@ -16,6 +14,7 @@ pub mod state_store;
 mod error;
 
 mod ledger;
+mod role;
 mod save_state;
 mod secret;
 mod secret_source;
@@ -24,8 +23,11 @@ mod secret_source;
 mod state_machine_test;
 
 pub use self::{
+    alice::Alice,
+    bob::Bob,
     error::Error,
     ledger::{ExtractSecret, FundTransaction, Ledger, RedeemTransaction, RefundTransaction},
+    role::*,
     save_state::SaveState,
     secret::{FromErr, RandomnessSource, Secret, SecretHash},
     secret_source::*,
